Move ukraine_alarm coordinator to separate module
